+++ pkg/scalers/metrics_api_scaler.go
@@ -49,7 +49,7 @@ type metricsAPIScalerMetadata struct {
 	UnsafeSsl                         bool            `keda:"name=unsafeSsl,order=triggerMetadata,default=false"`
 	AggregateFromKubeServiceEndpoints bool            `keda:"name=aggregateFromKubeServiceEndpoints,order=triggerMetadata,default=false"`
 	AggregationType                   AggregationType `keda:"name=aggregationType,order=triggerMetadata,default=average,enum=average;sum;max;min"`
-	timeout                           time.Duration   // custom HTTP client timeout
+	Timeout                           time.Duration   `keda:"name=timeout, order=triggerMetadata, optional"`
 	// Authentication parameters for connecting to the metrics API
 	MetricsAPIAuth *authentication.Config `keda:"optional"`
 
@@ -95,7 +95,13 @@ func NewMetricsAPIScaler(config *scalersconfig.ScalerConfig, kubeClient client.C
 		return nil, fmt.Errorf("error parsing metric API metadata: %w", err)
 	}
 
-	httpClient := kedautil.CreateHTTPClient(meta.timeout, meta.UnsafeSsl)
+	// handle HTTP client timeout
+	httpClientTimeout := config.GlobalHTTPTimeout
+	if meta.Timeout > 0 {
+		httpClientTimeout = time.Duration(meta.Timeout) * time.Millisecond
+	}
+
+	httpClient := kedautil.CreateHTTPClient(httpClientTimeout, meta.UnsafeSsl)
 
 	// Handle TLS configuration with authentication config
 	if meta.MetricsAPIAuth != nil && meta.MetricsAPIAuth.EnabledTLS() {
@@ -129,19 +135,10 @@ func parseMetricsAPIMetadata(config *scalersconfig.ScalerConfig) (*metricsAPISca
 	}
 
 	// Resolve HTTP client timeout
-	meta.timeout = config.GlobalHTTPTimeout
-	timeoutVal, err := getParameterFromConfig(config, "timeout", false)
-	if err == nil {
-		timeout, err := strconv.Atoi(timeoutVal)
-		if err != nil {
-			return nil, fmt.Errorf("unable to parse timeout: %w", err)
-		}
-
-		if timeout <= 0 {
-			return nil, fmt.Errorf("timeout must be greater than 0: %w", err)
-		}
-
-		meta.timeout = time.Duration(timeout) * time.Millisecond
+	meta.Timeout = config.GlobalHTTPTimeout
+	// validate the timeout
+	if meta.Timeout < 0 {
+		return nil, fmt.Errorf("timeout must be greater than 0: %d", meta.Timeout)
 	}
 
 	return meta, nil