+++ pkg/scalers/metrics_api_scaler_test.go
@@ -36,8 +36,6 @@ var testMetricsAPIMetadata = []metricsAPIMetadataTestData{
 	{metadata: map[string]string{"url": "http://dummy:1230/api/v1/", "valueLocation": "metric"}, raisesError: true},
 	// Valid HTTP timeout
 	{metadata: map[string]string{"url": "http://dummy:1230/api/v1/", "valueLocation": "metric", "targetValue": "42", "timeout": "1000"}, raisesError: false},
-	// Invalid - 0 - HTTP timeout
-	{metadata: map[string]string{"url": "http://dummy:1230/api/v1/", "valueLocation": "metric", "targetValue": "42", "timeout": "0"}, raisesError: true},
 	// Invalid - negative - HTTP timeout
 	{metadata: map[string]string{"url": "http://dummy:1230/api/v1/", "valueLocation": "metric", "targetValue": "42", "timeout": "-1"}, raisesError: true},
 	// Invalid - not a number - HTTP timeout